pact-python-master/requirements_dev.txt
Click>=2.0.0,<=6.7
Flask==1.0
configparser==3.5.0
codecov==2.0.5
coverage==4.3.4
enum34==1.1.6
flake8==3.2.1
mccabe==0.5.2
mock==2.0.0
nose==1.3.7
pycodestyle==2.2.0
pydocstyle==1.1.1
pyflakes==1.3.0
tox==2.5.0
tox-travis==0.8
wheel==0.24.0
pytest-xdist==1.31.0
pluggy<1.0,>=0.12
pytest-jira==0.3.15
pactman==2.25.0
flask==1.1.1